consolidate more clicktracking modules into tracking/util/
