--- src/actions/clickActionCreators.js
@@ -1,9 +1,0 @@
-export const CLICK_TRACK_ACTION = 'CLICK_TRACK';
-export const CLICK_TRACK_CLEAR_ACTION = 'CLICK_TRACK_CLEAR';
-
-export const click = clickData => ({
-	type: CLICK_TRACK_ACTION,
-	payload: clickData,
-});
-
-export const clearClick = () => ({ type: CLICK_TRACK_CLEAR_ACTION });
