+++ src/plugins/tracking/util/avro.test.js
@@ -1,6 +1,6 @@
 import avsc from 'avsc';
 import { getLogger } from '../activity';
-import { clickToClickRecord } from './clickTrackingReader';
+import { clickToClickRecord } from './clickReader';
 import * as avro from './avro';
 
 jest.mock('@google-cloud/pubsub', () => {
