+++ src/plugins/tracking/util/clickReader.js
