+++ src/plugins/tracking/util/clickReader.test.js
@@ -1,6 +1,4 @@
-import processClickTracking, {
-	clickCookieOptions,
-} from './clickTrackingReader';
+import processClickTracking, { clickCookieOptions } from './clickReader';
 
 describe('processClickTracking', () => {
 	const click = { lineage: '', coords: [1, 2] };
